+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# HEAD
+# Version 0.3.0
 
--   **Breaking change**: Implement the `demo` command and enable a nicer CLI interface [#5](https://github.com/simoneboscolo99/5K_JFS/pull/5)
--   Fixed an issue with the vertical order of the images [#4](https://github.com/simoneboscolo99/5K_JFS/pull/4)
+-   Implement a path tracer [#6](https://github.com/simoneboscolo99/5K_JFS/pull/6)
+
+# Version 0.2.0
+
+-   **Breaking change**: Implement the `demo` and `convert` commands and enable a nicer CLI interface [#5](https://github.com/simoneboscolo99/5K_JFS/pull/5)
+-   Fixed an issue with the vertical order of the images [#4](https://github.com/simoneboscolo99/5K_JFS/pull/4)  
 
 # Version 0.1.1